refactor(countervalues): move tracking-pair inference to live-common
`inferTrackingPairForAccounts`, its unresolved variant and the
`useTrackingPairForAccounts` hook that wraps them needed an `Account`,
which is the one thing the countervalues core had no business knowing
about. They now sit in `live-common/portfolio`, next to the portfolio
code they serve.

With them gone, `@ledgerhq/types-live` disappears from the package and
from its dependency list. The move had to wait for both countervalues
packages to become private: a published package cannot depend on a
private one, and the hook re-exported a function that had to land in
private `live-common`.

Consumer changes are import paths only.
