+++ apps/ledger-live-mobile/src/actions/general.ts
@@ -7,10 +7,8 @@ import {
 } from "@ledgerhq/live-common/account/ordering";
 import type { FlattenAccountsOptions } from "@ledgerhq/live-common/account/index";
 import type { TrackingPair } from "@ledgerhq/live-countervalues/types";
-import {
-  useCalculateCountervalueCallback as useCalculateCountervalueCallbackCommon,
-  useTrackingPairForAccounts,
-} from "@ledgerhq/live-countervalues-react";
+import { useCalculateCountervalueCallback as useCalculateCountervalueCallbackCommon } from "@ledgerhq/live-countervalues-react";
+import { useTrackingPairForAccounts } from "@ledgerhq/live-common/portfolio/useTrackingPairForAccounts";
 import { useDistribution as useLegacyDistribution } from "@ledgerhq/live-common/portfolio/portfolioReact";
 import {
   useAssetDistribution,
