+++ libs/ledger-live-common/src/portfolio/__tests__/useTrackingPairForAccounts.test.ts
@@ -1,0 +1,133 @@
+/**
+ * @jest-environment jsdom
+ */
+import { genAccount } from "@ledgerhq/ledger-wallet-framework/mocks/account";
+import type { FiatCurrency } from "@ledgerhq/ledger-wallet-framework/types";
+import type { TrackingPair } from "@ledgerhq/live-countervalues/types";
+import { renderHook, act } from "@testing-library/react";
+import { inferTrackingPairForAccounts } from "../trackingPairs";
+import { useTrackingPairForAccounts } from "../useTrackingPairForAccounts";
+
+const usd: FiatCurrency = {
+  type: "FiatCurrency",
+  name: "US Dollar",
+  ticker: "USD",
+  symbol: "$",
+  units: [{ name: "dollar", code: "USD", magnitude: 2, showAllDigits: true, prefixCode: true }],
+};
+const eur: FiatCurrency = {
+  type: "FiatCurrency",
+  name: "Euro",
+  ticker: "EUR",
+  symbol: "€",
+  units: [{ name: "euro", code: "EUR", magnitude: 2, showAllDigits: true, prefixCode: true }],
+};
+
+describe("useTrackingPairForAccounts", () => {
+  const accounts = Array(20)
+    .fill(null)
+    .map((_, i) => genAccount("test" + i));
+  const trackingPairs = inferTrackingPairForAccounts(accounts, usd);
+
+  test("it returns same tracking pairs as when using inferTrackingPairForAccounts", async () => {
+    const { result } = renderHook(() => useTrackingPairForAccounts(accounts, usd));
+    await act(async () => {
+      expect(result.current).toEqual(trackingPairs);
+    });
+  });
+
+  test("a re-render preserve the reference", async () => {
+    const { result, rerender } = renderHook(() => useTrackingPairForAccounts(accounts, usd));
+    let initial: TrackingPair[] | undefined;
+    await act(async () => {
+      initial = result.current;
+    });
+    rerender();
+    await act(async () => {
+      expect(result.current).toBe(initial);
+    });
+  });
+
+  test("a re-render preserve the reference even when accounts change", async () => {
+    const { result, rerender } = renderHook(() =>
+      useTrackingPairForAccounts(accounts.slice(0), usd),
+    );
+    let initial: TrackingPair[] | undefined;
+    await act(async () => {
+      initial = result.current;
+    });
+    rerender();
+    await act(async () => {
+      expect(result.current).toBe(initial);
+    });
+  });
+
+  test("when accounts appears, it properly converge to the trackingPairs", async () => {
+    const { result, rerender } = renderHook(added =>
+      useTrackingPairForAccounts(!added ? [] : accounts, usd),
+    );
+    await act(async () => {
+      expect(result.current).toEqual([]);
+    });
+    rerender(true);
+    await act(async () => {
+      expect(result.current).toEqual(trackingPairs);
+    });
+  });
+
+  test("when accounts changes fundamentally, pairs change", async () => {
+    const { result, rerender } = renderHook(empty =>
+      useTrackingPairForAccounts(empty ? [] : accounts, usd),
+    );
+    await act(async () => {
+      expect(result.current).toEqual(trackingPairs);
+    });
+    rerender(true);
+    await act(async () => {
+      expect(result.current).toEqual([]);
+    });
+  });
+
+  test("when currency changes, pairs change", async () => {
+    const { result, rerender } = renderHook(usesEur =>
+      useTrackingPairForAccounts(accounts, usesEur ? eur : usd),
+    );
+    await act(async () => {
+      expect(result.current).toEqual(trackingPairs);
+    });
+    rerender(true);
+    await act(async () => {
+      expect(result.current).not.toEqual(trackingPairs);
+    });
+  });
+
+  test("if accounts reorder, it doesn't change", async () => {
+    const reverse = accounts.slice(0).reverse();
+    const { result, rerender } = renderHook(rev =>
+      useTrackingPairForAccounts(rev ? reverse : accounts, usd),
+    );
+    let initial: TrackingPair[] | undefined;
+    await act(async () => {
+      initial = result.current;
+    });
+    rerender(true);
+    await act(async () => {
+      expect(result.current).toBe(initial);
+    });
+  });
+
+  test("if accounts doubles, it doesn't change", async () => {
+    const doubled = accounts.concat(accounts);
+    const { result, rerender } = renderHook(d =>
+      useTrackingPairForAccounts(d ? doubled : accounts, usd),
+    );
+    let initial: TrackingPair[] | undefined;
+    await act(async () => {
+      initial = result.current;
+    });
+    rerender(true);
+    await act(async () => {
+      expect(result.current).toBe(initial);
+    });
+  });
+});
