+++ libs/live-countervalues/package.json
@@ -18,7 +18,6 @@
   "types": "src/types.ts",
   "license": "Apache-2.0",
   "dependencies": {
-    "@ledgerhq/types-live": "workspace:*",
     "@ledgerhq/ledger-wallet-framework": "workspace:^",
     "@ledgerhq/logs": "catalog:",
     "@ledgerhq/live-network": "workspace:*",
