+++ libs/live-countervalues/src/logic.ts
@@ -1,9 +1,4 @@
 import { log } from "@ledgerhq/logs";
-import {
-  flattenAccounts,
-  getAccountCurrency,
-  isAccountEmpty,
-} from "@ledgerhq/ledger-wallet-framework/account/helpers";
 import { promiseAllBatched } from "@ledgerhq/live-promise";
 import type {
   CounterValuesState,
@@ -29,7 +24,6 @@ import {
   datapointRetention,
   inferCurrencyAPIID,
 } from "./helpers";
-import type { Account } from "@ledgerhq/types-live";
 import type { Currency } from "@ledgerhq/ledger-wallet-framework/types";
 import api from "./api";
 
@@ -122,33 +116,6 @@ export function importCountervalues(
     ),
     checkHolesOnNextLoad: true,
   };
-}
-
-// infer the tracking pair from user accounts to know which pairs are concerned
-export function inferTrackingPairForAccountsUnresolved(
-  accounts: Account[],
-  countervalue: Currency,
-): TrackingPair[] {
-  const yearAgo = new Date();
-  yearAgo.setFullYear(yearAgo.getFullYear() - 1);
-  yearAgo.setHours(0, 0, 0, 0);
-  return flattenAccounts(accounts)
-    .filter(a => !isAccountEmpty(a))
-    .map(account => {
-      const currency = getAccountCurrency(account);
-      return {
-        from: currency,
-        to: countervalue,
-        startDate: account.creationDate < yearAgo ? account.creationDate : yearAgo,
-      };
-    });
-}
-
-export function inferTrackingPairForAccounts(
-  accounts: Account[],
-  countervalue: Currency,
-): TrackingPair[] {
-  return resolveTrackingPairs(inferTrackingPairForAccountsUnresolved(accounts, countervalue));
 }
 
 /**
